% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableTerms.R
\name{availableTerms}
\alias{availableTerms}
\title{Extract the terms used under a given attributes/column}
\usage{
availableTerms(attributes, db = "cMD")
}
\arguments{
\item{attributes}{A character (1). Name of the attribute/column you want to 
extract the terms used under.}

\item{db}{A character(1). Currently, `cMD` (curatedMetagenomicData) is only
supported.}
}
\value{
A named list. Name of the list is the attribute/column name
}
\description{
Extract the terms used under a given attributes/column
}
\examples{
availableTerms("age_group")

}
