% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topNodes.R
\name{commonNodes}
\alias{commonNodes}
\title{Retrieves top nodes for a given ontology term map}
\usage{
commonNodes(ids, dbs)
}
\arguments{
\item{ids}{Character vector of term ids}

\item{dbs}{Character vector of corresponding ontology names. Single 
string also accepted if all terms share a single ontology.}
}
\value{
A Dataframe of chosen nodes including information on number of 
original terms covered
}
\description{
Retrieves top nodes for a given ontology term map
}
\examples{
ids <- c("NCIT:C983", "NCIT:C61612")
dbs <- c("NCIT", "NCIT")
commonNodes(ids = ids, dbs = dbs)


}
