% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateMetadata.R
\name{compress_metadata}
\alias{compress_metadata}
\title{Compresses expanded treatment columns to original format}
\usage{
compress_metadata(meta, idcols = ID_COLS, ccols = T_COLS, delim = "<;>")
}
\arguments{
\item{meta}{A data frame with expanded treatment columns}

\item{idcols}{Optional character vector of columns that identify single samples, defaults to standard ID columns}

\item{ccols}{Optional character vector of columns to compress if present, defaults to all treatment-related columns}

\item{delim}{Optional delimiter string, defaults to "<;>"}
}
\value{
A data frame where each sample gets a single row
}
\description{
Compresses expanded treatment columns to original format
}
