% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateMetadata.R
\name{compressMetadata}
\alias{compressMetadata}
\title{Compresses expanded treatment columns to original format}
\usage{
compressMetadata(meta, idcols = NULL, ccols = NULL, delim = "<;>")
}
\arguments{
\item{meta}{A data frame with expanded treatment columns.}

\item{idcols}{Optional. A character vector of columns that identify 
single samples. Defaults to standard ID columns.}

\item{ccols}{Optional. A character vector of columns to compress if 
present, defaults to all treatment-related columns.}

\item{delim}{Optional. A delimiter string. Default is \code{"<;>"}.}
}
\value{
A data frame where each sample gets a single row
}
\description{
Compresses expanded treatment columns to original format
}
