% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topNodes.R
\name{.displayNodes}
\alias{.displayNodes}
\title{Retrieves ontology terms and database information for given term ids}
\usage{
.displayNodes(onto, node)
}
\arguments{
\item{onto}{A character vector. Name(s) of ontologies that terms are from.}

\item{node}{A character vector of ontology term IDs.}
}
\value{
Dataframe of submitted term IDs, term names, and term ontologies
}
\description{
Retrieves ontology terms and database information for given term ids
}
