% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoFilter.R
\name{.getAllTargetForms}
\alias{.getAllTargetForms}
\title{Collect both labels and obo_ids related to the query term}
\usage{
.getAllTargetForms(query)
}
\arguments{
\item{query}{A character vector of terms or term ids. For example, 
`c("NCIT:C35025", "HP:0003003", "colon cancer")`.}
}
\value{
A character vector of all the related terms' label and obo_id.
}
\description{
This function can take multiple obo_id and labels, and return both 
labels/obo_ids exactly matching with the `query`.
}
