% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topNodes.R
\name{.getBottomDists}
\alias{.getBottomDists}
\title{Retrieves number of descendants of given term}
\usage{
.getBottomDists(onto, nodes)
}
\arguments{
\item{onto}{A character vector. Name(s) of ontologies that terms are from.}

\item{nodes}{A character vector of ontology term IDs.}
}
\value{
Dataframe of submitted terms and numbers of descendants
}
\description{
Retrieves number of descendants of given term
}
