% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMetadata.R
\name{.getOntoInfo}
\alias{.getOntoInfo}
\title{Collect both terms and ids related to the query}
\usage{
.getOntoInfo(query, return = NULL)
}
\description{
Collect both terms and ids related to the query
}
