% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topNodes.R
\name{.getTopDists}
\alias{.getTopDists}
\title{Retrieves number of ancestors of given term}
\usage{
.getTopDists(onto, nodes)
}
\arguments{
\item{onto}{A character vector. Name(s) of ontologies that terms are from.}

\item{nodes}{A character vector of ontology term IDs.}
}
\value{
Dataframe of submitted terms and numbers of ancestors
}
\description{
Retrieves number of ancestors of given term
}
