% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateMetadata.R
\name{expand_metadata}
\alias{expand_metadata}
\title{Expands treatment metadata}
\usage{
expand_metadata(meta, ecols = T_COLS, delim = "<;>")
}
\arguments{
\item{meta}{A data frame of metadata including all treatment-related columns}

\item{ecols}{Optional character vector of columns to expand if presesnt, defaults to all treatment-related columns}

\item{delim}{Optional delimiter string, defaults to "<;>"}
}
\value{
A data frame of metadata expanded so that each individual treatment has its own row
}
\description{
Expands treatment metadata
}
