% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMetadata.R
\name{filterMetadata}
\alias{filterMetadata}
\title{Filter metadata table by ontology term ids in specific attribute}
\usage{
filterMetadata(metaTb, feature, ids)
}
\arguments{
\item{metaTb}{Metadata table}

\item{feature}{A character vector (1). Column name to filter by}

\item{ids}{List of ids to filter by}
}
\value{
Metadata table filtered by provided ontology term ids in 
provided attribute
}
\description{
Filter metadata table by ontology term ids in specific attribute
}
\examples{
dir <- system.file("extdata", package="OmicsMLRepoR")
df <- read.csv(file.path(dir, "sample_metadata.csv"))
filterMetadata(df,
               "curated_disease",
               c("NCIT:C2855", "EFO:0000228"))
               
}
