% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMetadata.R
\name{findDistantRelatives}
\alias{findDistantRelatives}
\title{Find distant relatives}
\usage{
findDistantRelatives(pool, target)
}
\arguments{
\item{pool}{A named list of the ancestors for a given ontology term. The 
given ontology term is also included in the list and serves as a name of
the element.}

\item{target}{A character vector of ontology ids.}
}
\value{
A character vector of ontology term ids having any of `target` 
terms as their ancestor.
}
\description{
This function takes a list of ontology terms and their ancestors, where
the lowest descent serves as a name of the element, as a `pool`. Names 
of all the elements that have the `target` term in their ancestor list 
(i.e., sharing the ancestors) are returned.
}
\details{
This functionality is the base of searching metadata leveraging ontology, 
because a specific term can be used to query all the related, decendant
terms used in the metadata.
}
\examples{
ancestors <- list(
    "NCIT:C5490"=c("NCIT:C5490","NCIT:C4910","NCIT:C2955","NCIT:C4978"),
    "NCIT:C177680"=c("NCIT:C177680","NCIT:C4910","NCIT:C2955","NCIT:C4978"),
    "NCIT:C2955"=c("NCIT:C2955","NCIT:C4978","NCIT:C3141"))
targets <- "NCIT:C4910"
findDistantRelatives(pool = ancestors, target = targets)

}
