% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMetadata.R
\name{findRelated}
\alias{findRelated}
\title{Find base terms that are equivalent to or descendants of search results}
\usage{
findRelated(pool, target)
}
\arguments{
\item{pool}{A named list of all the ancestor terms for a given ontology 
term. The name of the element is the queried term for ancestor search.}

\item{target}{A character vector of ontology ids.}
}
\value{
A character vector of ontology term ids having any of `target` terms
as their ancestor.
}
\description{
Find base terms that are equivalent to or descendants of search results
}
\examples{
ancestors <- list("EFO:0000228" = c("EFO:0000228", "EFO:0000616"),
                  "NCIT:C2855" = c("NCIT:C2855", "NCIT:C3262"))
results <- c("NCIT:C3262", "EFO:0000616")
findRelated(pool = ancestors, target = results)

}
