% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topNodes.R
\name{findReps}
\alias{findReps}
\title{Chooses which ancestors cover all given original terms, prioritizing a low 
number of chosen nodes}
\usage{
findReps(onto, vecs)
}
\arguments{
\item{onto}{Character string; name of ontology}

\item{vecs}{List of character vectors of ancestors named by original node}
}
\value{
Dataframe of chosen nodes including information on number of 
original terms covered
}
\description{
Chooses which ancestors cover all given original terms, prioritizing a low 
number of chosen nodes
}
