% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byProperty.R
\name{get_children}
\alias{get_children}
\title{Get the parents of a requested ontology term}
\usage{
get_children(source, termId, plot = TRUE)
}
\description{
Get the parents of a requested ontology term
}
\examples{
get_children(source = "hp", termId ="HP:0025031")

}
