% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoTraverse.R
\name{get_parents}
\alias{get_parents}
\title{Get the parents of a requested ontology term}
\usage{
get_parents(source, termId, plot = TRUE)
}
\description{
Get the parents of a requested ontology term
}
\examples{
get_parents(source = "chebi", termId = "CHEBI:5262")
get_parents(source = "hp", termId ="HP:0025031")

}
