% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/byProperty.R
\name{get_siblings}
\alias{get_siblings}
\title{Get the siblings of a requested ontology term}
\usage{
get_siblings(source, termId, plot = TRUE)
}
\description{
Get the siblings of a requested ontology term
}
\examples{
get_siblings(source = "hp", termId = "HP:0003003")

}
