% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMetadata.R
\name{getAncestors}
\alias{getAncestors}
\title{Retrieve and save ancestors of a given ontology term}
\usage{
getAncestors(ids, dbs)
}
\arguments{
\item{ids}{A character vector of ontology term ids}

\item{dbs}{A character vector of corresponding ontologies, with the same
length of `ids`. Single string is
also accepted if all terms are from a single ontology.}
}
\value{
A named list of character vectors containing the ancestors of 
the original term. Name of each element is the original term the ancestors
are traced from.
}
\description{
Retrieve and save ancestors of a given ontology term
}
\examples{
ids <- c("EFO:0005856", "NCIT:C122177", "HP:0001541", 
         "NCIT:C28397", "EFO:0003914", "EFO:0003780")
dbs <- c("EFO", "NCIT", "HP", "NCIT", "EFO", "EFO")
getAncestors(ids, dbs)


}
