% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get data available through OmicsMLRepo}
\usage{
getData(
  metadata,
  dataType,
  cache_directory = get_default_cache_dir(),
  repository = TEMP_LOCAL_DIR
)
}
\arguments{
\item{metadata}{A data frame containing, at minimum, a `unique_id` column,
which corresponds to a `study_id:sample_id`. This can be obtained from
the \code{\link{getMetadata}} function.}

\item{dataType}{A character(1) describing a data type.}

\item{cache_directory}{Optional. A character(1). If provided, it should
indicate a local file path where any remotely accessed files should be
copied.}

\item{repository}{}
}
\description{
Get data available through OmicsMLRepo
}
