% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMetadata.R
\name{getMaps}
\alias{getMaps}
\title{Load ontology mapping files}
\usage{
getMaps(map_directory)
}
\arguments{
\item{map_directory}{A path to the directory containing .csv map files}
}
\value{
A list of maps as data frames
}
\description{
Load ontology mapping files
}
