% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Gets the harmonized metadata as a data frame}
\usage{
getMetadata(
  remote_url = DATABASE_URL,
  cache_directory = get_default_cache_dir(),
  use_cache = TRUE
)
}
\arguments{
\item{remote_url}{Optional character vector of length 1, containing the
location of metadata as an HTTP URL or a Google bucket address.}

\item{cache_directory}{Optional character vector of length 1. A file path
on your local system to a directory (not a file) that will be used to store
the metadata parquet file.}

\item{use_cache}{Logical. If \code{TRUE} (the default) and this function
has been called before with the same parameters, a cached reference to the
table will be returned. If \code{FALSE}, a new connection will be created.}
}
\value{
A lazy data.frame subclass containing the metadata. You can interact
with this object using most standard dplyr functions. For string matching,
it is recommended that you use \code{\link[stringr]{str_like}} to filter
character columns, as \code{\link[stringr]{str_match}} will not work.
}
\description{
Downloads a parquet database of the harmonized omics metadata deposited in
Bioconductor's ExperimentHub.
}
