% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topNodes.R
\name{getNodes}
\alias{getNodes}
\title{Retrieves all ancestors for supplied terms}
\usage{
getNodes(onto, terms)
}
\arguments{
\item{onto}{A character vector. Name(s) of ontologies that terms are from.}

\item{terms}{A character vector of ontology term IDs.}
}
\value{
A named list. Names of elements are original nodes (`terms`). 
Each element is a character vectors containing the ancestors of the 
element name (i.e., original terms provided).
}
\description{
Retrieves all ancestors for supplied terms
}
\examples{
terms <- c("NCIT:C25301", "NCIT:C29844", "NCIT:C29846", "NCIT:C29848")
getNodes(onto = "NCIT", terms = terms)


}
