% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarTerms.R
\name{getNumRelatives}
\alias{getNumRelatives}
\title{Get ancestors or descendants of a specified degree from the given term}
\usage{
getNumRelatives(term, relationship, num)
}
\arguments{
\item{term}{Character string; term id}

\item{relationship}{Character string; specifies type of desired result terms;
either "ancestors" or "descendants"}

\item{num}{Integer; how far from the given term to retrieve result terms}
}
\value{
A list of rols Term objects
}
\description{
Get ancestors or descendants of a specified degree from the given term
}
\examples{
getNumRelatives(term = "NCIT:C26902", relationship = "ancestors", num = 2)
getNumRelatives(term = "NCIT:C35708", relationship = "descendants", num = 1)

}
