% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOntoDesc.R
\name{getOntoDesc}
\alias{getOntoDesc}
\title{Look up the description of ontology terms}
\usage{
getOntoDesc(terms)
}
\arguments{
\item{terms}{A character vector containing ontology term ids to look up
their definition. Each ontology term id should be their official format
with their ontology symbol and term number separated by colon.}
}
\description{
Look up the description of ontology terms
}
\examples{
terms <- ontoTraverse(source = "hp", termId = "HP:0003003", plot = FALSE)
getOntoDesc(terms)

}
