% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{gs_file_size}
\alias{gs_file_size}
\title{Gets the file size of a number of remote files}
\usage{
gs_file_size(gss)
}
\arguments{
\item{gss}{A character vector containing Google bucket file location(s)}
}
\value{
The file size of each of the files pointed to by the provided
Google bucket address, in gigabytes, as double vector
}
\description{
Gets the file size of a number of remote files
}
\keyword{internal}
