% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HP}
\alias{HP}
\title{Human Phenotype Ontology index}
\format{
List of lists and vectors
}
\description{
Human Phenotype Ontology (HPO) comprising a list of 
lists/vectors of properties of HPO terms indexed by term ID
}
