% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateMetadata.R
\docType{data}
\name{ID_COLS}
\alias{ID_COLS}
\title{Character vector of ID columns}
\format{
An object of class \code{character} of length 4.
}
\usage{
ID_COLS
}
\description{
Character vector of ID columns
}
\keyword{datasets}
