% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{letters_only}
\alias{letters_only}
\title{Check that a character doesn't match any non-letter}
\usage{
letters_only(x)
}
\arguments{
\item{x}{A character(1).}
}
\description{
Check that a character doesn't match any non-letter
}
