% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_cbio}
\alias{mini_cbio}
\title{A subset of cBioPortalData metadata}
\format{
A data frame with 10 samples and 9 columns ('curation_id',
'acronym', 'acronym_ontology_term_id', 'sex', 'package', 'treatment_name',
'treatment_name_ontology_term_id', 'treatment_type', 
'treatment_type_ontology_term_id')
}
\usage{
mini_cbio
}
\description{
A subset of curated version of cBioPortal's clinical metadata.
}
\author{
Sehyun Oh \email{shbrief@gmail.com}
}
\keyword{data}
