% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_cmd}
\alias{mini_cmd}
\title{A subset of cMD metadata}
\format{
A data frame with 200 samples and 3 columns ('curation_id',
'hla', and 'package')
}
\usage{
mini_cmd
}
\description{
A subset of curated version of `sampleMetadata` from the 
curatedMetagenomicData (cMD, ver.3.8.0) package.
}
\author{
Sehyun Oh \email{shbrief@gmail.com}
}
\keyword{data}
