% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_cmd2}
\alias{mini_cmd2}
\title{A subset of cMD metadata}
\format{
A data frame with 200 samples and 7 columns ('curation_id', 
'target_condition', 'target_condition_ontology_term_id',
'pmid', 'disease', 'disease_ontology_term_id', 'package'). 
The two key exemplary attributes ('target_condition' and 'disease') 
selected here contain multiple values.
}
\usage{
mini_cmd2
}
\description{
A subset of curated version of `sampleMetadata` from the 
curatedMetagenomicData (cMD, ver.3.8.0) package.
}
\author{
Sehyun Oh \email{shbrief@gmail.com}
}
\keyword{data}
