% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mini_cmd3}
\alias{mini_cmd3}
\title{A subset of cMD metadata}
\format{
A data frame with 200 samples and 7 columns ('curation_id', 
'pmid', 'package', 'target_condition', 'feces_phenotype', 
'probing_pocket_depth', 'target_condition_ontology_term_id',
'feces_phenotype_ontology_term_id', 'probing_pocket_depth_ontology_term_id'). 
The 'target_condition' is multi-valued attribute, and 'feces_phenotype' and
'probing_pocket_depth' are composite attributes.
}
\usage{
mini_cmd3
}
\description{
A subset of curated version of `sampleMetadata` from the 
curatedMetagenomicData (cMD, ver.3.8.0) package.
}
\author{
Sehyun Oh \email{shbrief@gmail.com}
}
\keyword{data}
