% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{numbers_only}
\alias{numbers_only}
\title{Check that a character doesn't match any non-number}
\usage{
numbers_only(x)
}
\arguments{
\item{x}{A character(1).}
}
\description{
Check that a character doesn't match any non-number
}
