% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{data}
\name{OmicsMLRepo_URL}
\alias{OmicsMLRepo_URL}
\title{Base URL pointing to the count data at the current version}
\format{
An object of class \code{character} of length 1.
}
\usage{
OmicsMLRepo_URL
}
\description{
Base URL pointing to the count data at the current version
}
\keyword{datasets}
