% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoFilter.R
\name{onto_filter}
\alias{onto_filter}
\title{Keep rows that match a condition and}
\usage{
onto_filter(.data, col, query, delim = ";")
}
\arguments{
\item{.data}{A data frame}

\item{col}{A character (1). Column name to filter by.}

\item{query}{A character vector containing words or ids to be used in the
ontology search}

\item{delim}{A character (1) used to separate multiple values. Default ';'.}
}
\value{
Data frame filtered by provided queries along with their ontology
synonyms/ids in the specified column. Not case-sensitive.
}
\description{
Analog of Tidyverse 'filter' function that includes ontology synonyms and 
ids
}
\examples{
dir <- system.file("extdata", package="OmicsMLRepoR")
df <- read.csv(file.path(dir, "sample_metadata.csv"))
onto_filter(df, curated_disease, c("diabetes", "adenoma"))

}
