% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoBrowse.R
\name{ontoBrowse}
\alias{ontoBrowse}
\title{Browse ontology databases from ontology term or term id}
\usage{
ontoBrowse(term, ontoDB = NULL, rows = 10)
}
\arguments{
\item{term}{Character(1). The ontolgy term or term id you want 
to check its definition.}

\item{ontoDB}{A character vector containing the ontology databases you
want to look up the \code{term}. The other ontology databases you want to
look for the same/similar term as the input \code{term}.}

\item{rows}{Integer(1).The maximum number of terms to look up. Default is 10.}
}
\value{
Data
}
\description{
Browse ontology databases from ontology term or term id
}
\examples{
ontoBrowse("Sitagliptin", ontoDB = "Chebi")
ontoBrowse("Sitagliptin", ontoDB = c("Chebi", "apple"))
ontoBrowse("CHEBI:40237")

}
