% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMetadata.R
\name{ontoSearch}
\alias{ontoSearch}
\title{Search OLS for ontology terms with plain text input}
\usage{
ontoSearch(term, ontology = "", exact = FALSE, rows = 20)
}
\arguments{
\item{term}{A character (1) containing the search query, either a term or
ontology term id.}

\item{ontology}{A character (1) defining the ontology to be queried. 
Default is the empty character, to search all ontologies.}

\item{exact}{A logical (1) defining if OLS search is restricted to exact
matches. Defaults is `FALSE`.}

\item{rows}{An integer (1) defining the number of query returns. 
Default is 20L. Maximum number of values returned by the server is 1000.}
}
\value{
A character vector of matched ontology term ids
}
\description{
Search OLS for ontology terms with plain text input
}
\examples{
ontoSearch("plasma,membrane", ontology = "go")
ontoSearch("NCIT:C4872", rows = 5)
ontoSearch("Skin Infection", rows = 5)

}
