% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoTraverse.R
\name{ontoTraverse}
\alias{ontoTraverse}
\title{Extract ancestors/descendants of a term from ontology tree}
\usage{
ontoTraverse(source, termId, plot = TRUE, includeDefinition = FALSE)
}
\arguments{
\item{source}{Ontology database id. Currently available options are 
\code{NCIT, CHEBI, HP}.}

\item{termId}{Ontology term id started with \code{{ontology database id}:}, 
such as \code{NCIT:C15280}. Accept term itself in addition to term id??}

\item{plot}{Default is \code{TRUE}. If it is set to \code{FALSE}, this 
function will return a character vector of associate ontology terms.}
}
\value{
Ontology plot including queried term and its associated ancestors
and descendants terms. With the \code{plot=FALSE} argument, it will return
a character vector of associated ontology term ids.
}
\description{
Extract ancestors/descendants of a term from ontology tree
}
\examples{
ontoTraverse(source = "ncit", termId = "NCIT:C15280")
ontoTraverse(source = "chebi", termId = "CHEBI:5262")
ontoTraverse(source = "hp", termId = "HP:0003003")

}
