% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxoMap.R
\name{oxoMap}
\alias{oxoMap}
\title{Mapping input ontology terms and a target ontology through the OxO API}
\usage{
oxoMap(term_ids, target_ontology, mapping_distance = 1)
}
\arguments{
\item{term_ids}{A character vector of term ids to retrieve mappings for}

\item{target_ontology}{A character (1). The target ontology library to 
map the `term_ids` to.}

\item{mapping_distance}{Optional. Mapping distance as an integer, 1 
(default), 2, or 3.}
}
\value{
A list of of data frames with mapping results for each input term
}
\description{
Mapping input ontology terms and a target ontology through the OxO API
}
