% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{report_file_sizes}
\alias{report_file_sizes}
\title{Prints a message indicating the size of a download}
\usage{
report_file_sizes(urls)
}
\arguments{
\item{urls}{A character vector containing the location of metadata. All
should be HTTP URLs or Google bucket addresses - not accepting any mix of
them currently.}
}
\value{
`NULL`, invisibly
}
\description{
Prints a message indicating the size of a download
}
\keyword{internal}
