% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_metadata}
\alias{sample_metadata}
\title{sample_metadata}
\format{
A data frame with 4 rows and 7 columns
}
\usage{
sample_metadata
}
\description{
A small data table to demonstrating the data reshaping 
functions in OmicsMLRepoR.
}
\author{
Sehyun Oh \email{shbrief@gmail.com}
}
\keyword{data}
