% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMetadata.R
\name{saveAncestors}
\alias{saveAncestors}
\title{Wrapper to save ancestor info for all curated features within a metadata 
database. Currently, we have established curated features for 
curatedMetagenomicData (cMD) and cBioPortalData.}
\usage{
saveAncestors(curatedMetaTb, saveAs)
}
\arguments{
\item{curatedMetaTb}{A curated metadata table. Curated ontology terms should
be under the column name satisfying the patten, `^.*_ontology_term_id$`}

\item{saveAs}{A character (1). The file path for the returned list.}
}
\description{
Wrapper to save ancestor info for all curated features within a metadata 
database. Currently, we have established curated features for 
curatedMetagenomicData (cMD) and cBioPortalData.
}
