% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMetadata.R
\name{searchMetadata}
\alias{searchMetadata}
\title{Wrapper to easily search metadata with plain text}
\usage{
searchMetadata(
  term,
  metaTb,
  targetDB,
  feature = NULL,
  exact = FALSE,
  onto = "",
  rows = 20
)
}
\arguments{
\item{term}{A character (1) containing the search query.}

\item{metaTb}{A data frame. Metadata table containing `feature` where you
want to search the queried `term`.}

\item{targetDB}{A character (1) indicating ancestor file to access. 
Available options are `cMD` (curatedMetagenomicData) and `cBioPortalData`.}

\item{feature}{A character vector (1). Column name to filter by. Under the
default (`NULL`), all the columns using ontology terms are browsed.}

\item{exact}{A logical (1) defining if OLS search is restricted to exact
matches. Defaults is `FALSE`.}

\item{onto}{A character (1) defining the ontology to be queried for a 
provided `term`. Default is the empty character, to search all ontologies.}

\item{rows}{An integer (1) defining the number of query returns. 
Default is 20L. Maximum number of values returned by the server is 1000.}
}
\value{
A data frame which is the subset of the input metadata table 
(`metaTb`) filtered by the provided ontology term ids (`onto`) under 
the feature attribute (`feature`).
}
\description{
Wrapper to easily search metadata with plain text
}
\examples{
dir <- system.file("extdata", package = "OmicsMLRepoR")
metaTb <- read.csv(file.path(dir, "cMD_curated_metadata_all.csv"), header = TRUE)
searchMetadata(term = "cancer", 
               metaTb = metaTb, 
               targetDB = "cMD", 
               feature = "curated_disease")

}
