% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/similarTerms.R
\name{similarTerms}
\alias{similarTerms}
\title{Get same-level terms (cousins) of a term that are related through an ancestor
of a specified distance}
\usage{
similarTerms(term, dist = 1)
}
\arguments{
\item{term}{Character string; term id}

\item{dist}{Integer; max distance of joining ancestor; defaults to 1}
}
\value{
A list of rols Term objects
}
\description{
Get same-level terms (cousins) of a term that are related through an ancestor
of a specified distance
}
\examples{
similarTerms(term = "NCIT:C26902", dist = 1)

}
