% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sync_remote_file}
\alias{sync_remote_file}
\title{Synchronizes a single remote file with a local path}
\usage{
sync_remote_file(full_url, output_file, ...)
}
\arguments{
\item{full_url}{A character vector containing the location of metadata. All
should be HTTP URLs or Google bucket addresses - not accepting any mix of
them currently.}

\item{output_file}{File path to the cached data.}
}
\value{
`NULL`, invisibly
}
\description{
Synchronizes a single remote file with a local path
}
\keyword{internal}
