% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateMetadata.R
\docType{data}
\name{T_COLS}
\alias{T_COLS}
\title{Character vector of treatment-related columns (current but not complete)}
\format{
An object of class \code{character} of length 46.
}
\usage{
T_COLS
}
\description{
Character vector of treatment-related columns (current but not complete)
}
\keyword{datasets}
