% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\docType{data}
\name{TEMP_LOCAL_DIR}
\alias{TEMP_LOCAL_DIR}
\title{Temporary local directory containing exported empty, place-folder data files}
\format{
An object of class \code{character} of length 1.
}
\usage{
TEMP_LOCAL_DIR
}
\description{
Temporary local directory containing exported empty, place-folder data files
}
\keyword{datasets}
