% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoFilter.R
\name{tree_filter}
\alias{tree_filter}
\title{Analog of Tidyverse 'filter' function that includes ontology}
\usage{
tree_filter(.data, col, query, db, delim = ";")
}
\arguments{
\item{.data}{A data frame}

\item{col}{A character (1). Column name to filter by.}

\item{query}{A character vector containing words or ids to be used in the
ontology search}

\item{db}{A character (1) indicating ancestor file to access. Available
options are 'cMD' (curatedMetagenomicData) and 'cBioPortalData'.}

\item{delim}{A character (1) used to separate multiple values. Default ';'.}
}
\value{
Data frame filtered by provided queries along with child terms in the
specified column
}
\description{
Analog of Tidyverse 'filter' function that includes ontology
}
\examples{
meta <- getMetadata("cMD")
tree_filter(meta, curated_disease, c("pancreatic disease", "cancer"), "cMD")

}
