% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{unique_col_exists}
\alias{unique_col_exists}
\title{Check the existence of a column with all unique values}
\usage{
unique_col_exists(x)
}
\description{
Check the existence of a column with all unique values
}
