% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/availableTerms.R
\name{availableTerms}
\alias{availableTerms}
\title{Extract all the terms used in a quired attribute/column}
\usage{
availableTerms(attribute, db = "cMD")
}
\arguments{
\item{attribute}{A character (1). Name of the attribute/column you want to 
extract the terms used under.}

\item{db}{A character(1). Currently, `cMD` (curatedMetagenomicData) is only
supported.}
}
\value{
A data frame with two columns - existing values under the quried
attribute (`allowedvalues` column) and their ontology term id (`ontology`
column).
}
\description{
Extract all the terms used in a quired attribute/column
}
\examples{
availableTerms("age_group")
availableTerms("disease")

}
