% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchMetadata.R
\name{filterMetadata}
\alias{filterMetadata}
\title{Filter metadata table by ontology term ids}
\usage{
filterMetadata(metaTb, query, feature, delim = "<;>")
}
\arguments{
\item{metaTb}{Metadata table}

\item{query}{A character vector containing obo_ids to query}

\item{feature}{A character (1). Column name to filter by.}

\item{delim}{A character(1) used to separate multiple values. Default `<;>`.}
}
\value{
Metadata table filtered by provided ontology term ids in 
provided attribute
}
\description{
Filter metadata table by ontology term ids
}
\examples{
dir <- system.file("extdata", package="OmicsMLRepoR")
df <- read.csv(file.path(dir, "sample_metadata.csv"))
filterMetadata(df, c("NCIT:C2855", "EFO:0000228"), "curated_disease")
filterMetadata(df, c("diabetes", "EFO:0000228"), "curated_disease")

}
