% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topNodes.R
\name{findReps}
\alias{findReps}
\title{Chooses which ancestors cover all given original terms, prioritizing a low 
number of chosen nodes}
\usage{
findReps(onto, vecs)
}
\arguments{
\item{onto}{Character string; name of ontology}

\item{vecs}{List of character vectors of ancestors named by original node}
}
\value{
Dataframe of chosen nodes including information on number of 
original terms covered
}
\description{
Chooses which ancestors cover all given original terms, prioritizing a low 
number of chosen nodes
}
\examples{
onto <- "NCIT"
vecs <- list("NCIT:C983" = c("NCIT:C2846", "NCIT:C471", "NCIT:C1909",
"NCIT:C29711", "NCIT:C78276"), "NCIT:C61612" = c("NCIT:C1892", "NCIT:C1909",
"NCIT:C98234", "NCIT:C29711", "NCIT:C78276"))
findReps(onto = onto, vecs = vecs)


}
