% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{Get OmicsMLRepo data from curated metadata}
\usage{
getData(
  metadata,
  data_type,
  cache_directory = get_default_cache_dir(),
  repository = OmicsMLRepo_URL
)
}
\arguments{
\item{metadata}{A data frame containing, at minimum, one 'unique_id' column,
This input is a subset of the table obtained from \code{\link{getMetadata}}.}

\item{data_type}{A character(1) describing a data type. Unless all the 
queried samples should have data in this data type, it will give an error.}

\item{cache_directory}{Optional. A character(1). If provided, it should
indicate a local file path where any remotely accessed files should be
copied.}

\item{repository}{}
}
\description{
Get OmicsMLRepo data from curated metadata
}
