% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulateMetadataTable.R
\name{getLongMetaTb}
\alias{getLongMetaTb}
\title{Expands metadata with multiple values belong to a same attribute}
\usage{
getLongMetaTb(meta, targetCols = NULL, delim = NULL)
}
\arguments{
\item{meta}{A data frame. Metadata table containing all treatment-related 
columns}

\item{targetCols}{Optional. A character vector of column names to expand if 
present. Default is the name of all cBioPortal treatment-related columns.}

\item{delim}{Optional. A character (1) of a delimiter used to separate 
multiple values in the metadata table.}
}
\value{
A data frame of metadata expanded so that each individual treatment 
has its own row.
}
\description{
Different from the `getWideMetaTb` function, this function accepts multiple
target columns (`targetCols`). The difference comes from that the target
columns of this function are related, i.e., have the same number of 
elements separated by the `delim`, and the multiple values for each column
belongs to the same column/attribute/field, i.e., no additional column
name is required/provided.
}
\examples{
dir <- system.file("extdata", package = "OmicsMLRepoR")
meta <- read.csv(file.path(dir, "mini_cmd.csv"), header = TRUE)
lmeta <- getLongMetaTb(meta, "hla")
dim(meta) 
dim(lmeta) 

short_tb <- data.frame(
    ind = c("A", "B", "C", "D", "E"),
    aval = c("cat;dog", "chicken", "horse", "frog;pig", "snake"),
    cval = c(1, NA, 3, 4, 5),
    bval = c("red;blue", "yellow", "NA", "green;NA", "brown"))
    
getLongMetaTb(short_tb, c("aval", "bval"), delim = ";")

}
