% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMetadata.R
\name{getMetadata}
\alias{getMetadata}
\title{Download a curated metadata table}
\usage{
getMetadata(database = NULL, load = TRUE)
}
\arguments{
\item{database}{Name of the database to get the metadata from. Currently, 
there are two available options.
\itemize{
    \item \code{cMD} : metadata for curatedMetagenomicData
    \item \code{cBioPortalData} : metadata for cBioPortalData
}}

\item{load}{Default is \code{TRUE}. If it's set to \code{FALSE}, the 
metadata table is downloaded to cache but not loaded into memory.}
}
\value{
Curated metadata table or file cache location, if `load = FALSE`.
}
\description{
Download a curated metadata table
}
\examples{
cmd <- getMetadata("cMD")

}
