% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoTraverse.R
\name{ontoTraverse}
\alias{ontoTraverse}
\title{Extract both ancestors and descendants of an ontology term}
\usage{
ontoTraverse(term, returnDescription = FALSE)
}
\arguments{
\item{term}{Ontology term id ('obo_id'), such as \code{NCIT:C15280}.}

\item{returnDescription}{Default is `FALSE`. If it is set to `TRUE`, the
returned value includes description of the term and other information.}
}
\value{
A character vector of ancestors, self, and descendants ontology
term ids. If `returnDescription = TRUE`, it returns a tibble containig
details (including description) of the related ontology term ids.
}
\description{
Extract both ancestors and descendants of an ontology term
}
\examples{
ontoTraverse(term = "NCIT:C15280")
ontoTraverse(term = "CHEBI:5262")
ontoTraverse(term = "HP:0011793")

}
