% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoTraverse.R
\name{ontoTraverse}
\alias{ontoTraverse}
\title{Extract both ancestors and descendants of a term from ontology tree}
\usage{
ontoTraverse(ontoTermId, includeDefinition = FALSE)
}
\arguments{
\item{ontoTermId}{Ontology term id started with the ontology prefix and id
separated by colon, such as \code{NCIT:C15280}.}

\item{plot}{Under the default (\code{FALSE}), this function returns a 
character vector of associate ontology terms. If it is set to \code{TRUE}, 
it will plot the ontology tree of the returned term. (Currently, not 
supported.)}
}
\value{
Ontology plot including queried term and its associated ancestors
and descendants terms. With the \code{plot=FALSE} argument, it will return
a character vector of associated ontology term ids.
}
\description{
Extract both ancestors and descendants of a term from ontology tree
}
\examples{
ontoTraverse(ontoTermId = "NCIT:C15280")
ontoTraverse(ontoTermId = "CHEBI:5262")
ontoTraverse(ontoTermId = "HP:0011793")

}
