% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoTreePlot.R
\name{ontoTreePlot}
\alias{ontoTreePlot}
\title{Plot ontology tree}
\usage{
ontoTreePlot(term)
}
\arguments{
\item{term}{A character (1). Ontology term id (obo_id)}
}
\value{
A ontology tree plot. All the terms used in the output plot are 
ancestors of the queried term, so the queried term is the tip.
}
\description{
Plot ontology tree
}
\examples{
ontoTreePlot("NCIT:C2852")

}
