% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spreadMeta.R
\name{spreadMeta}
\alias{spreadMeta}
\title{Manipulate metadata into a spread format}
\usage{
spreadMeta(meta, targetCol)
}
\arguments{
\item{meta}{A data frame.}

\item{targetCol}{A character (1). The column name to expand if present.}
}
\value{
A metadata table where the `targetCol` is updated to a long form (if it
is multi-valued attribute) or spread into multiple columns (if it is 
composite attribute).
}
\description{
This function speads gout multiple values per attribute (‘multi-valued’ 
attribute) or multiple features under a single generic attribute 
(‘composite’ attribute). Multi-valued attribute will return 'long' table, 
while composite attribute will return 'wide' table.
}
